+++ src/gemseo_benchmark/results/history_item.py
@@ -27,21 +27,20 @@ class HistoryItem:
 
     def __init__(
         self,
-        # TODO: API BREAK: rename argument 'objective_value' into 'performance_measure'.
-        objective_value: float,
+        performance_measure: float,
         infeasibility_measure: float,
         n_unsatisfied_constraints: int | None = None,
     ) -> None:
         """
         Args:
-            objective_value: The performance measure of the item.
+            performance_measure: The performance measure of the item.
             infeasibility_measure: The infeasibility measure of the item.
             n_unsatisfied_constraints: The number of unsatisfied constraints of the
                 item.
                 If ``None``, it will be set to 0 if the infeasibility measure is zero,
                 and if the infeasibility measure is positive it will be set to None.
         """  # noqa: D205, D212, D415
-        self.__performance_measure = objective_value
+        self.__performance_measure = performance_measure
         (
             self.__infeas_measure,
             self.__n_unsatisfied_constraints,
@@ -96,9 +95,8 @@ class HistoryItem:
 
         return infeasibility_measure, n_unsatisfied_constraints
 
-    # TODO: API BREAK: rename property 'objective_value' into 'performance_measure'.
     @property
-    def objective_value(self) -> float:
+    def performance_measure(self) -> float:
         """The performance measure of the history item."""
         return self.__performance_measure
 
@@ -117,7 +115,7 @@ class HistoryItem:
         return self.__n_unsatisfied_constraints
 
     def __repr__(self) -> str:
-        return str((self.objective_value, self.infeasibility_measure))
+        return str((self.performance_measure, self.infeasibility_measure))
 
     def __eq__(self, other: HistoryItem) -> bool:
         """Compare the history item with another one for equality.
@@ -130,7 +128,7 @@ class HistoryItem:
         """
         return (
             self.__infeas_measure == other.__infeas_measure
-            and self.objective_value == other.objective_value
+            and self.performance_measure == other.performance_measure
         )
 
     def __lt__(self, other: HistoryItem) -> bool:
@@ -144,7 +142,7 @@ class HistoryItem:
         """
         return self.__infeas_measure < other.__infeas_measure or (
             self.__infeas_measure == other.__infeas_measure
-            and self.objective_value < other.objective_value
+            and self.performance_measure < other.performance_measure
         )
 
     def __le__(self, other: HistoryItem) -> bool: